Move Schemas to Schema.
