+++ src/Query/QueryBuilder.php
@@ -19,8 +19,8 @@ use Yiisoft\Db\Expression\ExpressionBuilderInterface;
 use Yiisoft\Db\Expression\ExpressionInterface;
 use Yiisoft\Db\Query\Conditions\HashCondition;
 use Yiisoft\Db\Query\Conditions\SimpleCondition;
-use Yiisoft\Db\Schemas\ColumnSchemaBuilder;
-use Yiisoft\Db\Schemas\Schema;
+use Yiisoft\Db\Schema\ColumnSchemaBuilder;
+use Yiisoft\Db\Schema\Schema;
 use Yiisoft\Db\Pdo\PdoValue;
 use Yiisoft\Db\Pdo\PdoValueBuilder;
 use Yiisoft\Strings\StringHelper;
